refactor: remove 'near-places-cards.tsx'
