+++ project/src/components/offer-сards/offer-cards.tsx
@@ -1,71 +1,36 @@
-import { useState } from 'react';
-import { City, Offer, Offers } from 'src/types/types';
+import { Offer, Offers } from 'src/types/types';
 import OfferCard from 'src/components/offer-card/offer-card';
-import GeoMap from 'src/components/geo-map/geo-map';
-import { getMultipleOfPlaceWord } from 'src/utils/utils';
-
-type ActiveOffer = Offer | null;
 
 type OfferCardsProps = {
+  children?: JSX.Element;
+  className: string;
+  header: string;
   offers: Offers;
-  offersCount: number;
-  currentCity: City;
+  onActive: (item: Offer) => void;
+  onBlur: () => void;
 }
 
 function OfferCards(props: OfferCardsProps): JSX.Element {
-  const [ hoveredOffer, setHoveredOffer ] = useState<ActiveOffer>(null);
-
   return (
-    <div className="cities">
-      <div className="cities__places-container container">
-        <section className="cities__places places">
-          <h2 className="visually-hidden">Places</h2>
-          <b className="places__found">
-            {`
-              ${props.offersCount}
-              ${getMultipleOfPlaceWord(props.offersCount)}
-              to stay in ${props.currentCity.name}
-            `}
-          </b>
-          <form className="places__sorting" action="#" method="get">
-            <span className="places__sorting-caption">Sort by</span>
-            <span className="places__sorting-type" tabIndex={0}>
-              Popular
-              <svg className="places__sorting-arrow" width="7" height="4">
-                <use xlinkHref="#icon-arrow-select"></use>
-              </svg>
-            </span>
-            <ul className="places__options places__options--custom">{ /* FIXME 'places__options--opened' */}
-              <li className="places__option places__option--active" tabIndex={0}>Popular</li>
-              <li className="places__option" tabIndex={0}>Price: low to high</li>
-              <li className="places__option" tabIndex={0}>Price: high to low</li>
-              <li className="places__option" tabIndex={0}>Top rated first</li>
-            </ul>
-          </form>
-          <div className="cities__places-list places__list tabs__content">
-            {
-              props.offers.map((offer: Offer): JSX.Element => (
-                <OfferCard
-                  className='cities__card'
-                  key={offer.id}
-                  offer={offer}
-                  onActive={() => setHoveredOffer(offer)}
-                  onBlur={() => setHoveredOffer(null)}
-                />
-              ))
-            }
-          </div>
-        </section>
-        <div className="cities__right-section">
-          <GeoMap
-            activeOffer={hoveredOffer}
-            className={props.offers.length <= 0 ? 'cities__map' : ''}
-            currentCity={props.currentCity}
-            offers={props.offers}
-          />
-        </div>
+    <section className={`${props.className} places`}>
+      <h2 className="visually-hidden">${props.header}</h2>
+
+      {props.children}
+
+      <div className="cities__places-list places__list tabs__content">
+        {
+          props.offers.map((offer: Offer): JSX.Element => (
+            <OfferCard
+              className='cities__card'
+              key={offer.id}
+              offer={offer}
+              onActive={() => props.onActive(offer)}
+              onBlur={() => props.onBlur()}
+            />
+          ))
+        }
       </div>
-    </div>
+    </section>
   );
 }
